Adaptation + Export Interface + External API Client
