+++ .prettierrc.js
@@ -1,4 +1,4 @@
 module.exports = {
-  singleQuote: true,
+  singleQuote: false,
   semi: true
 }
