+++ src/interfaces/ContainerInterface.ts
@@ -1,0 +1,3 @@
+export interface ContainerInterface {
+    deployment: string;
+  }
