+++ src/interfaces/ImageInterface.ts
@@ -1,0 +1,3 @@
+export interface ImageInterface {
+    deployment: string;
+  }
