+++ src/interfaces/Network/NetworkConnectInterface.ts
@@ -1,0 +1,3 @@
+export interface NetworkConnectInterface {
+    container: string;
+  }
