+++ src/interfaces/Network/NetworkCreateInterface.ts
@@ -1,0 +1,5 @@
+export interface NetworkCreateInterface {
+    name: string;
+    description: string;
+    subnet: string;
+  }
