--- src/interfaces/deployment.interface.ts
@@ -1,11 +1,0 @@
-export default interface DeploymentInterface
-{
-    name: string;
-    description: string;
-    image: string;
-    ports: string[];
-    mounts: string[];
-    environment: any[];
-    labels: any[];
-    links: any[];
-}
